Change repo of Neo4j Graph View
I'd like to change the name of the repo of Neo4j Graph View from `semantic-markdown-converter` to `obsidian-neo4j-graph-view`. 
The old name isn't representative of the contents of the repo anymore (the Python library called semantic-markdown-converter will be deprecated in the next version). 

Will everything keep working if the repo is changed like this? (Ie, all installations and stats are kept, just the link changes?)