Add Kindle Highlights plugin
Adding my plugin which is very much still in an alpha release — only tested on Mac at the moment.

Confirming that the package can be successfully downloaded by Obsidian's community plugins flow is akin to confirming that I have a continuous release pipeline confirmed in place. I plan to rapidly iterate on the plugin once this is merged or tested by other members in the community.