Add Request
A plugin that lets you perform fetch requests from obsidian. It appends the selected text to one of a list of configured urls and replaces the selected text with the response.

The full URL is included for the repo becuase it is hosted on gitlab. I made the personal decision to switch to gitlab after they non-consentually scraped public github repos for a machine learning tool (1). If it's an issue, I can create a github mirror, but considering Obsidian has a tech-savvy dev team and git itself is universal, maybe it should be possible to include other git hosts as well.

1. https://www.theverge.com/2021/7/7/22561180/github-copilot-legal-copyright-fair-use-public-code