Added weekly review notes linker plugin
Weekly review notes linker plugin is a fork of the weekly review plugin. It alters the main functionality to make a weekly note that links all the notes created in a week into one file. This is helpful if you use periodic notes. For periodic notes set your naming scheme for weekly notes to: Weekly Review - 202x-xx-xx