Update community-plugins.json
I know that there is already a plugin like that, but i was never satisfied with their keywords volume and also no support for russian language. So i took my chances and tried my best to compile my first obsidian plugin - and using rust for faster emoji search was a neat way to go.