Replace unmaintained plugin at-people
This plugin seems to have been unmaintained for the past 2 years. There are several open PR's and issues on the repo where no response was provided by the author. Github user hExPY has also tried to email the author but did not get a response. This fork aims to address the raised issues and restore active development